pyproject.toml
[tool.poetry]
name = "homeharvest"
version = "0.8.18"
description = "Real estate scraping library"
authors = ["Zachary Hampton <zachary@bunsly.com>", "Cullen Watson <cullen@bunsly.com>"]
homepage = "https://github.com/ZacharyHampton/HomeHarvest"
readme = "README.md"
packages = [
    { include = "homeharvest" },
    { include = "agent" },
]

[tool.poetry.dependencies]
python = ">=3.9"
requests = "^2.32.4"
pandas = "^2.3.1"
pydantic = "^2.11.7"
tenacity = "^9.1.2"
python-dotenv = "^1.0.1"
openai = "^1.108.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
